+++ vdf/Cargo.toml
@@ -9,6 +9,4 @@ description = "Pie19 VDF (https://eprint.iacr.org/2018/627.pdf)."
 elapsed = "0.1.2"
 lazy_static = "1.3.0"
 rug = "1.3.0"
-serialization = { path = "../serialization" }
-serialization_derive = { path = "../serialization_derive" }
 sha2 = "0.9.1"
